Extracted classes involved in external communication to "Models" folder
