+++ Frontend/src/app/Models/MapDescription.ts
