--- Frontend/src/app/hex/imap.ts
@@ -1,10 +1,0 @@
-import { ITile } from './itile';
-
-export class IMap {
-    id: number;
-    name: string;
-    yMin: number;
-    xMins: Array<number>;
-    xWidths: Array<number>;
-    tiles: Array<Array<ITile>>;
-}
